Move positionsSnapshot sync-schema to separate file
